sep the csv writer
