+++ config.ini
@@ -35,7 +35,8 @@ modules = nlisim.modules.molecules.Molecules
           nlisim.modules.pneumocyte.Pneumocyte
           nlisim.modules.liver.Liver
           nlisim.modules.state_output.StateOutput
-          nlisim.modules.visualization.Visualization
+          nlisim.modules.csv_writer.CSVWriter
+          #nlisim.modules.visualization.Visualization
 
 [state_output]
 # save the simulation state every 2 simulation minutes
@@ -250,9 +251,11 @@ hep_slope = 0.7793 # [il6] > 1e-11
 il6_threshold = 1.372243e-10  # 1e-11 # 5e-14 #molar (threshold to activate liver)
 threshold_log_hep = -8
 
-[visualization]
+[csv_writer]
 time_step = 2 # min
-csv_output = true
+
+[visualization]
+time_step = 10 # min
 # vtk_type: STRUCTURED_POINTS, STRUCTURED_GRID, RECTILINEAR_GRID, UNSTRUCTURED_GRID, POLY_DATA
 visual_variables = [
                    {
